% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexi_svd.R
\name{flexi_svd}
\alias{flexi_svd}
\title{Hyper-flexible SVD.}
\usage{
flexi_svd(Y, prop_row = 0.5, prop_col = 0.5, itermax = 500,
  max_rank = NULL)
}
\arguments{
\item{Y}{A matrix of data. There can be no missing values (yet).}

\item{prop_row}{The proportion of rows in Y11.}

\item{prop_col}{The proportion of columns in Y11.}

\item{itermax}{The maximum number of repetitions to run.}

\item{max_rank}{The maximum rank to consider.}
}
\description{
Hyper-flexible SVD.
}
\author{
David Gerard
}
