% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_svd.R
\name{js_svd}
\alias{js_svd}
\title{Josse-Sardy svd.}
\usage{
js_svd(Y, prop_row = 0.5, prop_col = 0.5, itermax = 500)
}
\arguments{
\item{Y}{A matrix of data. There can be no missing values (yet).}

\item{prop_row}{The proportion of rows in Y11.}

\item{prop_col}{The proportion of columns in Y11.}

\item{itermax}{The maximum number of repetitions to run.}
}
\description{
Josse-Sardy svd.
}
\references{
Josse, Julie, and Sylvain Sardy. "Adaptive shrinkage of singular values." Statistics and Computing 26.3 (2016): 715-724.
}
\author{
David Gerard
}
