% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot.truncsvd}
\alias{plot.truncsvd}
\title{Plots output from \code{\link{trunc_svd}}.}
\usage{
\method{plot}{truncsvd}(x, ...)
}
\arguments{
\item{x}{The output of \code{\link{trunc_svd}}.}

\item{...}{Not used.}
}
\description{
Plots output from \code{\link{trunc_svd}}.
}
\seealso{
\code{\link{trunc_svd}}
}
\author{
David Gerard
}
