% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_svd.R
\name{soft_obj}
\alias{soft_obj}
\title{The objective function to minimize in \code{\link{soft_svd}}.}
\usage{
soft_obj(lambda, Y22, Y21V, UtY12, d)
}
\arguments{
\item{lambda}{The tuning parameter.}

\item{Y22}{The lower right matrix.}

\item{Y21V}{The estimate of U2}

\item{UtY12}{The estimate of V2^t}

\item{d}{The singular values of Y11.}
}
\value{
The sums of squares difference from Y22 of the estimate of Y22.
}
\description{
The objective function to minimize in \code{\link{soft_svd}}.
}
\seealso{
\code{\link{soft_svd}}
}
\author{
David Gerard
}
