% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft_svd.R
\name{soft_svd}
\alias{soft_svd}
\title{Soft-thresholding SVD estimator.}
\usage{
soft_svd(Y, prop_row = 0.5, prop_col = 0.5, itermax = 500)
}
\arguments{
\item{Y}{A matrix of data. There can be no missing values (yet).}

\item{prop_row}{The proportion of rows in Y11.}

\item{prop_col}{The proportion of columns in Y11.}

\item{itermax}{The maximum number of repetitions to run.}
}
\value{
A list with the following elements
\itemize{
  \item{"rank_dist"}{The distribution of ranks.}
  \item{"d"}{The empirical singular values.}
  \item{"dhat"}{The estimated singular values}
  \item{"dlower"}{The lower bound for the 95\% interval}
  \item{"dupper"}{The upper bound for the 95\% interval}
  \item{"lambda_samp"}{The samples of the shrinkage parameter.}
}
}
\description{
Soft-thresholding SVD estimator.
}
\author{
David Gerard
}
