% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc_svd.R
\name{trunc_svd}
\alias{trunc_svd}
\title{Truncated SVD with uncertainty measures.}
\usage{
trunc_svd(Y, prop_row = 0.5, prop_col = 0.5, max_rank = NULL,
  itermax = 500)
}
\arguments{
\item{Y}{A matrix of data. There can be no missing values (yet).}

\item{prop_row}{The proportion of rows in Y11.}

\item{prop_col}{The proportion of columns in Y11.}

\item{max_rank}{The maximum rank to consider.}

\item{itermax}{The maximum number of repetitions to run.}
}
\description{
Truncated SVD with uncertainty measures.
}
\author{
David Gerard
}
